Moved plotting functions in a sub-directory for code clarity.
