+++ fast_pedago/plots/functions/aircraft_side_view.py
@@ -3,7 +3,7 @@ import plotly.graph_objects as go
 
 from fastoad.io import VariableIO
 
-from .plot_constants import (
+from ..plot_constants import (
     COLORS,
     NACELLE_POSITION,
     HT_HEIGHT,
