+++ fast_pedago/plots/functions/polar_with_lift_to_drag_ratio.py
@@ -6,7 +6,7 @@ import plotly.graph_objects as go
 
 from fastoad.io import VariableIO
 
-from .plot_constants import COLORS
+from ..plot_constants import COLORS
 
 
 def _polar_with_L_R_ratio_plot(
