+++ fast_pedago/plots/functions/stability_diagram.py
