+++ fast_pedago/plots/functions/static_margin.py
@@ -3,7 +3,7 @@ import plotly.graph_objects as go
 
 from fastoad.io import VariableIO
 
-from .plot_constants import COLORS
+from ..plot_constants import COLORS
 
 
 def _static_margin_plot(
