+++ fast_pedago/plots/functions/wing.py
