+++ fast_pedago/plots/plot_signatures.py
@@ -6,15 +6,17 @@ so they can be plotted using the same base function.
 
 import plotly.graph_objects as go
 
-from .aircraft_front_view import _aircraft_front_view_plot
-from .aircraft_side_view import _aircraft_side_view_plot
-from .aircraft_top_view import _aircraft_top_view_plot
-from .flaps_and_slats import _flaps_and_slats_plot
-from .simplified_payload_range import _simplified_payload_range_plot
-from .stability_diagram import _stability_diagram_plot
-from .wing import _wing_plot
-from .polar_with_lift_to_drag_ratio import _polar_with_L_R_ratio_plot
-from .static_margin import _static_margin_plot
+from .functions import (
+    _aircraft_front_view_plot,
+    _aircraft_side_view_plot,
+    _aircraft_top_view_plot,
+    _flaps_and_slats_plot,
+    _polar_with_L_R_ratio_plot,
+    _simplified_payload_range_plot,
+    _stability_diagram_plot,
+    _static_margin_plot,
+    _wing_plot,
+)
 
 import fastoad.api as oad
